telegram-bot/requirements.txt
openai>=1.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
supabase>=2.0.0
pyjwt>=2.8.0
cryptography>=41.0.0
python-telegram-bot[job-queue]>=20.7
flask>=3.0.0
praw>=7.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
langchain-community>=0.0.20
pydantic>=2.0.0
